+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDKConfiguration.java
@@ -35,9 +35,9 @@ class SDKConfiguration {
 	}};
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
-	public String sdkVersion = "1.35.1";
-	public String genVersion = "2.143.2";
-	public String userAgent = "speakeasy-sdk/java 1.35.1 2.143.2 0.1.0 <no value>";
+	public String sdkVersion = "1.35.2";
+	public String genVersion = "2.144.7";
+	public String userAgent = "speakeasy-sdk/java 1.35.2 2.144.7 0.1.0 <no value>";
 	public Map<String, Map<String, Map<String, Object>>> globals = new HashMap<String, Map<String, Map<String, Object>>>(){{
 		put("parameters", new HashMap<String, Map<String, Object>>());
 	}};