+++ python-client-sdk/src/sdk/sdkconfiguration.py
@@ -38,9 +38,9 @@ class SDKConfiguration:
     globals: dict[str, dict[str, dict[str, Any]]] = field(default_factory=dict)
     language: str = 'python'
     openapi_doc_version: str = '0.1.0'
-    sdk_version: str = '1.36.1'
-    gen_version: str = '2.143.2'
-    user_agent: str = 'speakeasy-sdk/python 1.36.1 2.143.2 0.1.0 openapi'
+    sdk_version: str = '1.36.2'
+    gen_version: str = '2.144.7'
+    user_agent: str = 'speakeasy-sdk/python 1.36.2 2.144.7 0.1.0 openapi'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: