+++ python-client-sdk/tests/test_telemetry.py
@@ -17,7 +17,7 @@ def test_telemetry_user_agent_get():
     assert res is not None
     assert res.status_code == 200
     assert res.res is not None
-    assert res.res.headers['X-Speakeasy-User-Agent'] == 'speakeasy-sdk/python 1.36.1 2.143.2 0.1.0 openapi'
+    assert res.res.headers['X-Speakeasy-User-Agent'] == 'speakeasy-sdk/python 1.36.2 2.144.7 0.1.0 openapi'
 
 
 def test_telemetry_speakeasy_user_agent_get():
@@ -33,4 +33,4 @@ def test_telemetry_speakeasy_user_agent_get():
     assert res.status_code == 200
     assert res.res is not None
     assert res.res.headers['User-Agent'] == user_agent
-    assert res.res.headers['X-Speakeasy-User-Agent'] == 'speakeasy-sdk/python 1.36.1 2.143.2 0.1.0 openapi'
+    assert res.res.headers['X-Speakeasy-User-Agent'] == 'speakeasy-sdk/python 1.36.2 2.144.7 0.1.0 openapi'