+++ typescript-client-sdk/src/sdk/retries.ts
@@ -52,7 +52,7 @@ export class Retries {
         headers["x-speakeasy-user-agent"] = this.sdkConfiguration.userAgent;
 
         const globalRetryConfig = this.sdkConfiguration.retryConfig;
-        let retryConfig: any = retries;
+        let retryConfig: utils.RetryConfig | undefined = retries;
         if (!retryConfig) {
             if (globalRetryConfig) {
                 retryConfig = globalRetryConfig;